+++ components/src/core/ChannelValue/ChannelValue.tsx
@@ -68,18 +68,16 @@ const Root = styled(Box)<Pick<ChannelValueProps, 'fontSize' | 'color'>>(({ fontS
     [`& .${channelValueClasses.text}`]: { fontSize: 'inherit', lineHeight: 'inherit', letterSpacing: 0 },
 }));
 
-const IconSpan = styled('span', {
-    name: 'channel-value',
-    slot: 'icon',
-})(() => ({
+const IconSpan = styled(
+    'span',
+    {}
+)(() => ({
     marginRight: '0.35em',
     display: 'inline',
     fontSize: 'inherit',
 }));
 
 const Unit = styled(Typography, {
-    name: 'channel-value',
-    slot: 'units',
     shouldForwardProp: (prop) => prop !== 'isSuffix',
 })<TypographyProps & { isSuffix: boolean }>(({ isSuffix }) => ({
     fontWeight: 300,
@@ -89,8 +87,6 @@ const Unit = styled(Typography, {
 }));
 
 const Value = styled(Typography, {
-    name: 'channel-value',
-    slot: 'value',
     shouldForwardProp: (prop) => prop !== 'isPrefix',
 })<TypographyProps & { isPrefix: boolean }>(({ isPrefix }) => ({
     fontWeight: 600,