+++ components/src/core/InfoListItem/InfoListItemStyledComponents.tsx
@@ -12,8 +12,6 @@ import Chevron from '@mui/icons-material/ChevronRight';
 import Typography, { TypographyProps } from '@mui/material/Typography';
 
 export const Root = styled(ListItem, {
-    name: 'info-list-item',
-    slot: 'root',
     shouldForwardProp: (prop) =>
         prop !== 'iconColor' &&
         prop !== 'iconColor' &&
@@ -29,7 +27,7 @@ export const Root = styled(ListItem, {
     >
 >(({ onClick, backgroundColor, wrapSubtitle, wrapTitle, wrapInfo, dense, ripple, theme }) => {
     const isWrapEnabled = (): boolean => wrapSubtitle || wrapTitle || wrapInfo;
-    const getHeight = (): string => (dense ? `3.25rem` : `4.5rem`);
+    const getHeight = (): string => (dense ? `3.5rem` : `4.5rem`);
 
     let isCssColor = true;
     try {
@@ -54,6 +52,18 @@ export const Root = styled(ListItem, {
             outline: 'none',
         },
         padding: onClick && ripple ? 0 : undefined,
+        paddingTop: isWrapEnabled() ? '0' : undefined,
+        paddingBottom: isWrapEnabled() ? '0' : undefined,
+        '& .MuiListItemButton-root': {
+            minHeight: isWrapEnabled() ? getHeight() : 'initial',
+            height: !isWrapEnabled() ? getHeight() : 'auto',
+            paddingTop: isWrapEnabled() ? '0' : undefined,
+            paddingBottom: isWrapEnabled() ? '0' : undefined,
+        },
+        '& .MuiListItemText-dense': {
+            marginTop: 0,
+            marginBottom: 0,
+        },
     };
 });
 
@@ -63,8 +73,6 @@ export const InfoListItemContentContainer = styled(ListItemButton)(() => ({
 }));
 
 export const StatusStripe = styled(Box, {
-    name: 'info-list-item',
-    slot: 'statusStripe',
     shouldForwardProp: (prop) => prop !== 'statusColor',
 })<Pick<InfoListItemProps, 'statusColor'>>(({ statusColor }) => ({
     position: 'absolute',
@@ -85,8 +93,6 @@ export const InfoListItemDivider = styled(Divider)<Pick<InfoListItemProps, 'divi
 }));
 
 export const Icon = styled(Avatar, {
-    name: 'info-list-item',
-    slot: 'icon',
     shouldForwardProp: (prop) =>
         prop !== 'statusColor' &&
         prop !== 'iconColor' &&
@@ -141,8 +147,6 @@ export const Icon = styled(Avatar, {
 );
 
 export const InfoListItemText = styled(ListItemText, {
-    name: 'info-list-item',
-    slot: 'listItemText',
     shouldForwardProp: (prop) => prop !== 'leftComponent',
 })<Pick<InfoListItemProps, 'leftComponent'>>(({ leftComponent, theme }) => ({
     // we have to specify both here because the auto-swap from JSS isn't smart enough to do it when we use a function
@@ -151,8 +155,6 @@ export const InfoListItemText = styled(ListItemText, {
 }));
 
 export const Subtitle = styled(Typography, {
-    name: 'info-list-item',
-    slot: 'subtitle',
     shouldForwardProp: (prop) => prop !== 'fontColor',
 })<Pick<InfoListItemProps & TypographyProps & BoxProps, 'fontColor' | 'component' | 'noWrap'>>(
     ({ fontColor, theme }) => ({
@@ -163,8 +165,6 @@ export const Subtitle = styled(Typography, {
 );
 
 export const Info = styled(Typography, {
-    name: 'info-list-item',
-    slot: 'info',
     shouldForwardProp: (prop) => prop !== 'fontColor',
 })<Pick<InfoListItemProps & TypographyProps & BoxProps, 'fontColor' | 'component' | 'noWrap'>>(
     ({ fontColor, theme }) => ({
@@ -174,10 +174,10 @@ export const Info = styled(Typography, {
     })
 );
 
-export const RightComponent = styled(Box, {
-    name: 'info-list-item',
-    slot: 'rightComponent',
-})(({ theme }) => ({
+export const RightComponent = styled(
+    Box,
+    {}
+)(({ theme }) => ({
     flex: '0 0 auto',
     marginLeft: theme.spacing(2),
     display: 'flex',
@@ -185,8 +185,6 @@ export const RightComponent = styled(Box, {
 }));
 
 export const InfoListItemChevron = styled(Chevron, {
-    name: 'info-list-item',
-    slot: 'chevron',
     shouldForwardProp: (prop) => prop !== 'chevronColor',
 })<Pick<InfoListItemProps, 'chevronColor'>>(({ chevronColor, theme }) => ({
     color: chevronColor ? chevronColor : theme.palette.text.secondary,