+++ components/src/core/ToolbarMenu/ToolbarMenu.tsx
@@ -43,10 +43,10 @@ export type ToolbarMenuCompGroup = {
     title?: string;
 };
 
-const Root = styled(Typography, {
-    name: 'toolbar-menu',
-    slot: 'root',
-})(() => ({
+const Root = styled(
+    Typography,
+    {}
+)(() => ({
     cursor: 'pointer',
     display: 'flex',
     alignItems: 'center',
@@ -54,35 +54,35 @@ const Root = styled(Typography, {
     [`&.${toolbarMenuClasses.cursorPointer}`]: { cursor: 'pointer ' },
 }));
 
-const DropDownArrow = styled(ArrowDropDown, {
-    name: 'toolbar-menu',
-    slot: 'drop-down-arrow',
-})(({ theme }) => ({
+const DropDownArrow = styled(
+    ArrowDropDown,
+    {}
+)(({ theme }) => ({
     marginLeft: theme.spacing(0.5),
     [`&.${toolbarMenuClasses.rotatedDropdownArrow}`]: { transform: 'rotate(180deg)' },
 }));
 
-const ToolbarMenuIcon = styled(Box, {
-    name: 'toolbar-menu',
-    slot: 'icon',
-})(({ theme }) => ({
+const ToolbarMenuIcon = styled(
+    Box,
+    {}
+)(({ theme }) => ({
     marginRight: theme.spacing(1),
     display: 'inline-flex',
     fontSize: 'inherit',
 }));
 
-const ToolbarMenuLabel = styled(Box, {
-    name: 'toolbar-menu',
-    slot: 'label',
-})(() => ({
+const ToolbarMenuLabel = styled(
+    Box,
+    {}
+)(() => ({
     textOverflow: 'ellipsis',
     overflow: 'hidden',
 }));
 
-const ToolbarMenuNavGroups = styled(Box, {
-    name: 'toolbar-menu',
-    slot: 'nav-groups',
-})(() => ({
+const ToolbarMenuNavGroups = styled(
+    Box,
+    {}
+)(() => ({
     '&:active, &:focus': {
         outline: 'none',
     },