+++ docs/src/index.tsx
@@ -13,13 +13,15 @@ import { BrowserRouter } from 'react-router-dom';
 import { createTheme, ThemeProvider, StyledEngineProvider } from '@mui/material/styles';
 import CssBaseline from '@mui/material/CssBaseline';
 import * as BLUIThemes from '@brightlayer-ui/react-themes';
+import useMediaQuery from '@mui/material/useMediaQuery';
 import '@brightlayer-ui/react-themes/open-sans';
 import '@fontsource/roboto';
 import '@fontsource/roboto-mono';
 import { App } from './App';
 import reportWebVitals from './reportWebVitals';
 import './index.css';
-import { store } from './redux/store';
+import { store, RootState } from './redux/store';
+import { useAppSelector } from './redux/hooks';
 import { MDXProvider } from '@mdx-js/react';
 import { componentsMap } from './__configuration__/markdownMapping';
 import { GoogleAnalyticsWrapper } from './router/GoogleAnalyticsWrapper';
@@ -30,11 +32,10 @@ import 'prismjs/plugins/line-numbers/prism-line-numbers.js';
 import 'prismjs/plugins/line-highlight/prism-line-highlight.js';
 
 // google analytics
-import ReactGA from 'react-ga';
-import { gaID } from './ga.js';
+import ReactGA from 'react-ga4';
 import { ScrollToTop } from './router/ScrollToTop';
-if (gaID) {
-    ReactGA.initialize(gaID);
+if (process.env.REACT_APP_GAID) {
+    ReactGA.initialize(process.env.REACT_APP_GAID);
 }
 
 // Brightlayer UI Icon font
@@ -46,20 +47,38 @@ if (!container) throw new Error('Root Element was not found in the DOM');
 const root = ReactDOMClient.createRoot(container);
 const basename = process.env.PUBLIC_URL || '/';
 
+const ThemedApp = (): JSX.Element => {
+    const siteTheme = useAppSelector((state: RootState) => state.appState.siteTheme);
+    const prefersDarkMode = useMediaQuery('(prefers-color-scheme: dark)');
+    let theme = BLUIThemes.blue;
+    if (siteTheme === 'dark' || (siteTheme === 'system' && prefersDarkMode)) {
+        theme = BLUIThemes.blueDark;
+    }
+
+    // force an update
+    const MemoThemedApp = React.useCallback(
+        () => (
+            <ThemeProvider theme={createTheme(theme)}>
+                <MDXProvider components={componentsMap as any}>
+                    <App />
+                </MDXProvider>
+            </ThemeProvider>
+        ),
+        [siteTheme, prefersDarkMode]
+    );
+    return <MemoThemedApp />;
+};
+
 root.render(
     <StyledEngineProvider injectFirst>
-        <ThemeProvider theme={createTheme(BLUIThemes.blue)}>
-            <BrowserRouter basename={basename}>
-                <ScrollToTop />
-                <GoogleAnalyticsWrapper />
-                <CssBaseline />
-                <Provider store={store}>
-                    <MDXProvider components={componentsMap as any}>
-                        <App />
-                    </MDXProvider>
-                </Provider>
-            </BrowserRouter>
-        </ThemeProvider>
+        <BrowserRouter basename={basename}>
+            <ScrollToTop />
+            <GoogleAnalyticsWrapper />
+            <CssBaseline />
+            <Provider store={store}>
+                <ThemedApp />
+            </Provider>
+        </BrowserRouter>
     </StyledEngineProvider>
 );