+++ docs/src/router/GoogleAnalyticsWrapper.tsx
@@ -1,11 +1,11 @@
 import React, { useEffect } from 'react';
 import { useLocation } from 'react-router-dom';
-import ReactGA from 'react-ga';
+import ReactGA from 'react-ga4';
 
 export const GoogleAnalyticsWrapper: React.FC = () => {
     const location = useLocation();
     useEffect(() => {
-        ReactGA.pageview(window.location.pathname + window.location.search);
+        ReactGA.send(window.location.pathname + window.location.search);
     }, [location.pathname, location.search]);
 
     return null;