+++ docs/src/shared/constants.ts
@@ -8,6 +8,11 @@ export const PAGE_WIDTH = {
     REGULAR: 766 + PADDING * 2, // text
 };
 
+export const CTA_BUTTON = {
+    WIDTH: 320 - PADDING * 2,
+    HEIGHT: 154,
+};
+
 export const REGULAR_WIDTH_STYLE = {
     marginLeft: 'auto',
     marginRight: 'auto',